Create TestingFibonacciJSWithClasses.html
Ok, so, now I went with a whole different aproach.
It was really fucking hard doind stuff in a functional mindset of programming
and, since JS is a multi-paradigm programming language, I came to test those object oriented programming stuff, ya know?
So, I started working with classes and methods instead of functions.

The classes I created so far:

-class FibonaccisArrayMaths: 
 a whole class dedicated to the maths behind the array and the Fibonacci thing. It is pretty much what I did on Java, but, this time, on JS.

-class  initialPositions:
Okay, I'm really proud of that one, so, I noticed that is a pattern of repetition for the next square. say, you start with a square, and then you put the next square on the left. then the squares after this particular second square that you've drawn will be: a square at the bottom, then a square at the right, then a square at the top, then a square at the left again.
basicaly: left, bottom, right, top
This sequence repeats itself, and after a whole afternoon spent on how to translate that to java, i found a way using the "%" keyword idea!
Basicaly, if (x - 1 / 4) is a natural number, it means it's left. If (x - 2/4) is a natural number, it means it's bottom and so on.

so I translated this to Javascript, and then, atribuitted to it a value, which will make easier to code
1 = left
2 =  bottom
3 = right
4 = top.

class positions:
-okay, here's the bad boy I'm working on. I had a really bad approach, and i'm trying a new one. I'll dedicate a whole new file based on that, so I can put on it later.